moved some sampleAdditionalInfo fields to sampleTissue
