+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/LibraryAdditionalInfo.java
@@ -16,14 +16,6 @@ public interface LibraryAdditionalInfo {
 
   void setLibrary(Library library);
 
-  TissueOrigin getTissueOrigin();
-
-  void setTissueOrigin(TissueOrigin tissueOrigin);
-
-  TissueType getTissueType();
-
-  void setTissueType(TissueType tissueType);
-
   Long getGroupId();
 
   void setGroupId(Long groupId);
