+++ migration/src/main/java/uk/ac/bbsrc/tgac/miso/migration/destination/MisoServiceManager.java
@@ -501,6 +501,7 @@ public class MisoServiceManager {
     svc.setTissueOriginDao(tissueOriginDao);
     svc.setTissueTypeDao(tissueTypeDao);
     svc.setSampleDao(sampleDao);
+    svc.setLabService(labService);
     setSampleService(svc);
   }
   
@@ -1180,6 +1181,7 @@ public class MisoServiceManager {
   
   private void updateLabServiceDependencies() {
     if (sampleAdditionalInfoService != null) sampleAdditionalInfoService.setLabService(labService);
+    if (sampleService != null) sampleService.setLabService(labService);
   }
 
   public HibernateLabDao getLabDao() {
