+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/SampleIdentityDto.java
@@ -44,4 +44,11 @@ public class SampleIdentityDto extends SampleAdditionalInfoDto {
     this.donorSex = donorSex;
   }
 
+  @Override
+  public String toString() {
+    return "SampleIdentityDto [internalName=" + internalName
+        + ", externalName=" + externalName + ", donorSex=" + donorSex
+        + ", super=" + super.toString() + "]";
+  }
+
 }
